feat: Support calling the OpenAI API directly (#85)
## Description

Support for going through the OpenAI API directly in addition to the
existing Azure support (API and secrets). Also moves to the `model`
notation as the `engine` param is deprecated and failed when hitting the
OpenAI API directly.

Updates the README with more clarity on the order of operations

## Testing

TODO

## Additional context

I dont work at microsoft nor have an Azure account so I'm not sure how
the engine vs. model distinction works on a custom Azure OpenAI endpoint
🤔

---------

Co-authored-by: Daniel Ciborowski <dciborow@microsoft.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>